Implemented module.export from initializer
